+++ src/pipeline/pipeline.mli
@@ -23,8 +23,12 @@ val stable_compatible : string -> string -> unit Diag.result
 val generate_idl : string list -> Idllib.Syntax.prog Diag.result
 
 val initial_stat_env : Scope.scope
-val chase_imports : parse_fn -> Scope.scope -> Resolve_import.resolved_imports ->
-  (Syntax.lib list * Scope.scope) Diag.result
+val chase_imports
+  :  stable_baseline_post:Type.field list option
+  -> parse_fn
+  -> Scope.scope
+  -> Resolve_import.resolved_imports
+  -> (Syntax.lib list * Scope.scope) Diag.result
 
 val run_files           : string list -> unit option
 val run_stdin_from_file : string list -> string -> Mo_values.Value.value option